+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -70,7 +70,6 @@ object Constants {
         const val INTELLIJ_PLATFORM_LOCAL = "intellijPlatformLocal"
         const val INTELLIJ_PLATFORM_DEPENDENCY = "intellijPlatformDependency"
         const val INTELLIJ_PLATFORM_PLUGIN_DEPENDENCY = "intellijPlatformPluginDependency"
-        const val INTELLIJ_PLATFORM_PLUGIN_DEPENDENCY_COLLECTOR = "intellijPlatformPluginDependencyCollector"
         const val INTELLIJ_PLATFORM_PLUGIN_LOCAL = "intellijPlatformPluginLocal"
         const val INTELLIJ_PLATFORM_PLUGIN_MODULE = "intellijPlatformPluginModule"
         const val INTELLIJ_PLATFORM_PLUGIN = "intellijPlatformPlugin"