+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Introduce configurations for IntelliJ Platform test plugins, dependencies, and bundled modules along with `testPlugin`, `testBundledPlugin`, and `testBundledModule` dependency helpers
+- Support different `kotlinx.coroutines` JavaAgent FQNs by checking for the class presence in the IntelliJ Platform classpath
 
 ## [2.4.0] - 2025-03-20